pyproject.toml
[tool.poetry]
name = "companies"
version = "0.1.0"
description = ""
authors = ["Ivan Kovtun <ivan.kovtun@capgemini.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
psycopg2 = "^2.9.9"
sqlmodel = "^0.0.18"
python-dotenv = "^1.0.1"
alembic = "^1.13.1"


[tool.poetry.group.dev.dependencies]
pydantic-settings = "^2.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
